Create Frequency B (Correlator)
Description of the problem I faced: 

This is the code I implemented in order to design the synchroniser based on correlator principle, After having convolved the received sequence with the preamble, I try to figure out the start of the signal by finding the peaks. The result of the convolution shows a clear pattern. However, I couldn't relate the pattern to the bit sequence. The pattern is as following: 
peak at pos. 3636, after exactly 896  symbols (896 = 640 + 2*128) there is another peak.  (the next peak is at 7148). 

Furthermore, I am also not sure about my decoding procedure. Although it succeeded in decoding the signal corresponding to frequency A, it could be somehow wrong for this case.